+++ interfaces/taihe/src/defined_form_taihe.cpp
@@ -46,12 +46,12 @@ void SystemDefinedFormTaihe::SetDetails(
     this->value_->SetDetails(udmfDetails);
 }
 
-double SystemDefinedFormTaihe::GetFormId()
+int32_t SystemDefinedFormTaihe::GetFormId()
 {
     return this->value_->GetFormId();
 }
 
-void SystemDefinedFormTaihe::SetFormId(double formId)
+void SystemDefinedFormTaihe::SetFormId(int32_t formId)
 {
     this->value_->SetFormId(formId);
 }
